Create stumble_g.py
Optimization is often defined as a minimum, and in our examples using multiple parameters a function root would serve as this minimum. For a search of any given root a start point would be chosen and assumed to not be within our tolerance. An initial search radius would be set and a random point within this radius, with the initial point as the center, would be selected. A comparison between the two points to determine which was better (the lesser value) would determine whether the new point or the original would act as the center for the next iteration of the optimum search. If no new center is selected after numerous different comparisons it can be assumed that the amount of points within the radius that are worse than the initial point is greater than the amount of points that are better. Our probability of selecting a good point is low so we decrease our search radius in hopes to eliminate the undesirable points. This is a predicted symptom of our arriving closer to our optimum as the likelihood of any point being greater than the point we currently are observing increases as the point we observe becomes less than its predecessor. The steps stated can be repeated until our search radius is within our tolerance meaning that the point currently observed is within our tolerance and we have found our optimum.